Correccion de errores, Creacion de MAIN, pruebas sobre DOCKER Y la DB, para ello se cambio el nombre de la DB a Integrador

+++ integrador-1/src/main/java/integrador_1/dao/FacturaProductoDAO.java
@@ -6,16 +6,16 @@ import java.util.List;
 import java.util.Optional;
 
 public interface FacturaProductoDAO {
-    FacturaProducto findById(Long id);
+    FacturaProducto findById(int id);
     List<FacturaProducto> findAll();
     //y si un producto nunca se facturó?
-    Optional<List<FacturaProducto>> findByProducto(Long idProducto);
+    Optional<List<FacturaProducto>> findByProducto(int idProducto);
 
     void create(FacturaProducto fp);
     void update(FacturaProducto fp);
     //idFactura es PK, pero la FK se tiene que poder updatear
-    void updateProducto(Long id, Long idProducto);
-    void delete(Long id);
+    void updateProducto(int id, int idProducto);
+    void delete(int id);
 
     void deleteAll();
 }
